move notification logic to command
