+++ src/Api/Vault/Controllers/SecurityTaskController.cs
@@ -2,16 +2,10 @@
 using Bit.Api.Vault.Models.Request;
 using Bit.Api.Vault.Models.Response;
 using Bit.Core;
-using Bit.Core.Enums;
-using Bit.Core.NotificationCenter.Commands.Interfaces;
-using Bit.Core.NotificationCenter.Entities;
-using Bit.Core.NotificationCenter.Enums;
-using Bit.Core.Repositories;
 using Bit.Core.Services;
 using Bit.Core.Utilities;
 using Bit.Core.Vault.Commands.Interfaces;
 using Bit.Core.Vault.Enums;
-using Bit.Core.Vault.Models.Data;
 using Bit.Core.Vault.Queries;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -28,10 +22,7 @@ public class SecurityTaskController : Controller
     private readonly IMarkTaskAsCompleteCommand _markTaskAsCompleteCommand;
     private readonly IGetTasksForOrganizationQuery _getTasksForOrganizationQuery;
     private readonly ICreateManyTasksCommand _createManyTasksCommand;
-    private readonly IGetSecurityTasksNotificationDetailsQuery _getSecurityTasksNotificationDetailsQuery;
-    private readonly IOrganizationRepository _organizationRepository;
-    private readonly IMailService _mailService;
-    private readonly ICreateNotificationCommand _createNotificationCommand;
+    private readonly ICreateManyTaskNotificationsCommand _createManyTaskNotificationsCommand;
 
     public SecurityTaskController(
         IUserService userService,
@@ -39,20 +30,14 @@ public class SecurityTaskController : Controller
         IMarkTaskAsCompleteCommand markTaskAsCompleteCommand,
         IGetTasksForOrganizationQuery getTasksForOrganizationQuery,
         ICreateManyTasksCommand createManyTasksCommand,
-        IGetSecurityTasksNotificationDetailsQuery getSecurityTasksNotificationDetailsQuery,
-        IOrganizationRepository organizationRepository,
-        IMailService mailService,
-        ICreateNotificationCommand createNotificationCommand)
+        ICreateManyTaskNotificationsCommand createManyTaskNotificationsCommand)
     {
         _userService = userService;
         _getTaskDetailsForUserQuery = getTaskDetailsForUserQuery;
         _markTaskAsCompleteCommand = markTaskAsCompleteCommand;
         _getTasksForOrganizationQuery = getTasksForOrganizationQuery;
         _createManyTasksCommand = createManyTasksCommand;
-        _getSecurityTasksNotificationDetailsQuery = getSecurityTasksNotificationDetailsQuery;
-        _organizationRepository = organizationRepository;
-        _mailService = mailService;
-        _createNotificationCommand = createNotificationCommand;
+        _createManyTaskNotificationsCommand = createManyTaskNotificationsCommand;
     }
 
     /// <summary>
@@ -106,35 +91,7 @@ public class SecurityTaskController : Controller
     {
         var securityTasks = await _createManyTasksCommand.CreateAsync(orgId, model.Tasks);
 
-        var securityTaskCiphers = await _getSecurityTasksNotificationDetailsQuery.GetNotificationDetailsByManyIds(orgId, securityTasks);
-
-        // Get the number of tasks for each user
-        var userTaskCount = securityTaskCiphers.GroupBy(x => x.UserId).Select(x => new
-        UserSecurityTasksCount
-        {
-            UserId = x.Key,
-            Email = x.First().Email,
-            TaskCount = x.Count()
-        }).ToList();
-
-        var organization = await _organizationRepository.GetByIdAsync(orgId);
-        await _mailService.SendBulkSecurityTaskNotificationsAsync(organization.Name, userTaskCount);
-
-        foreach (var userSecurityTaskCipher in securityTaskCiphers)
-        {
-            // Get the associated security task for the cipher
-            var securityTask = securityTasks.First(x => x.CipherId == userSecurityTaskCipher.CipherId);
-            // Create a notification for the user with the associated task
-            var notification = new Notification
-            {
-                UserId = userSecurityTaskCipher.UserId,
-                OrganizationId = orgId,
-                Priority = Priority.Informational,
-                ClientType = ClientType.Browser,
-                TaskId = securityTask.Id
-            };
-            await _createNotificationCommand.CreateAsync(notification);
-        }
+        await _createManyTaskNotificationsCommand.CreateAsync(orgId, securityTasks);
 
         var response = securityTasks.Select(x => new SecurityTasksResponseModel(x)).ToList();
         return new ListResponseModel<SecurityTasksResponseModel>(response);
